Storage APIs can throw arbitrary exceptions.  Be resilient to them.
